added AuthorizationService and rest authorizing mechanism

+++ src/main/java/com/jaewa/timesheet/controller/WorkdayController.java
@@ -3,7 +3,11 @@ package com.jaewa.timesheet.controller;
 import com.jaewa.timesheet.controller.dto.WorkdayDto;
 import com.jaewa.timesheet.controller.mapper.WorkdayMapper;
 import com.jaewa.timesheet.exception.IncoherentDataException;
+import com.jaewa.timesheet.exception.UnauthorizedException;
+import com.jaewa.timesheet.model.ApplicationUser;
 import com.jaewa.timesheet.model.Workday;
+import com.jaewa.timesheet.service.ApplicationUserService;
+import com.jaewa.timesheet.service.AuthorizationService;
 import com.jaewa.timesheet.service.WorkdayService;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -17,17 +21,28 @@ public class WorkdayController {
 
     private final WorkdayService workdayService;
 
+    private final ApplicationUserService applicationUserService;
+
     private final WorkdayMapper workdayMapper;
 
-    public WorkdayController(WorkdayService workdayService, WorkdayMapper workdayMapper) {
+    public WorkdayController(WorkdayService workdayService, ApplicationUserService applicationUserService, WorkdayMapper workdayMapper) {
         this.workdayService = workdayService;
+        this.applicationUserService = applicationUserService;
         this.workdayMapper = workdayMapper;
     }
 
     @GetMapping("/workday")
     public ResponseEntity<List<WorkdayDto>> getUserWorkDays(@RequestParam(value = "username") String username
-    ) {
-        //TODO only admin or owner user can access data
+    ) throws UnauthorizedException {
+
+        Optional<ApplicationUser> user = applicationUserService.getByUsername(username);
+
+        if (user.isEmpty()) {
+            return ResponseEntity.badRequest().build();
+        }
+
+        AuthorizationService.checkUserIsAuthorized(user.get().getId());
+
         return ResponseEntity.ok(
                 workdayService.findWorkdayByUser(username).stream()
                         .map(workdayMapper::toDto)
@@ -35,34 +50,35 @@ public class WorkdayController {
     }
 
     @GetMapping("/workday/{id}")
-    public ResponseEntity<WorkdayDto> getWorkDay(@PathVariable(value = "id") Long id) {
-        //TODO only admin or owner user can access data
-        Optional<Workday> result = workdayService.findById(id);
+    public ResponseEntity<WorkdayDto> getWorkDay(@PathVariable(value = "id") Long id) throws UnauthorizedException {
+        Optional<Workday> workday = workdayService.findById(id);
 
-        if (result.isEmpty()) {
+        if (workday.isEmpty()) {
             return ResponseEntity.badRequest().build();
         }
 
-        return ResponseEntity.ok(workdayMapper.toDto(result.get()));
+        AuthorizationService.checkUserIsAuthorized(workday.get().getApplicationUser().getId());
+
+        return ResponseEntity.ok(workdayMapper.toDto(workday.get()));
     }
 
     @PostMapping("/workday/new")
-    public ResponseEntity<WorkdayDto> addNewWorkday(@RequestBody WorkdayDto dto) throws IncoherentDataException {
-        //TODO only admin or owner user can add data
+    public ResponseEntity<WorkdayDto> addNewWorkday(@RequestBody WorkdayDto dto) throws IncoherentDataException, UnauthorizedException {
+        AuthorizationService.checkUserIsAuthorized(dto.getUsernameId());
         Workday result = workdayService.addNewWorkday(workdayMapper.toModel(dto));
         return ResponseEntity.ok(workdayMapper.toDto(result));
     }
 
     @PutMapping("/workday/edit")
-    public ResponseEntity<WorkdayDto> editWorkday(@RequestBody WorkdayDto dto) throws IncoherentDataException {
-        //TODO only admin or owner user can add data
+    public ResponseEntity<WorkdayDto> editWorkday(@RequestBody WorkdayDto dto) throws IncoherentDataException, UnauthorizedException {
+        AuthorizationService.checkUserIsAuthorized(dto.getUsernameId());
         Workday result = workdayService.editWorkday(workdayMapper.toModel(dto));
         return ResponseEntity.ok(workdayMapper.toDto(result));
     }
 
     @DeleteMapping("/workday/delete")
-    public ResponseEntity<String> deleteWorkday(@RequestBody WorkdayDto dto) {
-        //TODO only admin or owner user can add data
+    public ResponseEntity<String> deleteWorkday(@RequestBody WorkdayDto dto) throws UnauthorizedException {
+        AuthorizationService.checkUserIsAuthorized(dto.getUsernameId());
         workdayService.deleteWorkday(workdayMapper.toModel(dto));
         return ResponseEntity.ok().build();
     }
